pyproject.toml
[project]
name = "ragfit"
version = "1.2.0"
description = "Framework for enhancing LLMs for RAG tasks using fine-tuning."
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.10"
dependencies = [
    "bert-score>=0.3.13",
    "bitsandbytes>=0.42.0",
    "datasets>=2.16.1",
    "evaluate>=0.4.1",
    "hydra-core>=1.3.2",
    "nltk>=3.9",
    "openai>=1.23.3",
    "peft>=0.11.1",
    "pyyaml>=6.0.1",
    "rouge-score>=0.1.2",
    "sentence-transformers>=2.4.0",
    "sentencepiece>=0.2.0",
    "torch>=2.4.0",           # 降级到稳定版，2.4.0+ 支持大多数新特性
    "transformers>=4.45.0",    # 确保支持 Qwen3 系列架构
    "trl>=0.9.0",              # 升级 TRL 以解决你之前的 DataCollator 导入问题
    "wandb>=0.16.4",
]

[project.urls]
Homepage = "https://github.com/IntelLabs/RAG-FiT"
Documentation = "https://intellabs.github.io/RAG-FiT/"

[tool.setuptools]
packages = ["ragfit"]

[project.optional-dependencies]
deepeval = [
    "deepeval==0.21.73",
]
haystack = [
    "haystack-ai==2.3.1",
    "qdrant-haystack>=5.0.0",
]

